UpdateSemanticInfo event
