+++ src/Controls/samples/Controls.Sample/XamlApp.xaml.cs
@@ -17,7 +17,7 @@ namespace Maui.Controls.Sample
 
 			Debug.WriteLine($"The injected text service had a message: '{textService.GetText()}'");
 
-			MainPage = Services.GetRequiredService<Page>();
+			MainPage = new Pages.TapGestureGalleryPage();//Services.GetRequiredService<Page>();
 
 			RequestedThemeChanged += (sender, args) =>
 			{
@@ -25,6 +25,7 @@ namespace Maui.Controls.Sample
 				Debug.WriteLine($"Requested theme changed: {args.RequestedTheme}");
 			};
 		}
+
 
 		public IServiceProvider Services { get; }
 	}
