+++ src/Controls/src/Core/HandlerImpl/View.Impl.cs
@@ -26,5 +26,24 @@ namespace Microsoft.Maui.Controls
 
 			base.OnHandlerChangingCore(args);
 		}
+
+#if __ANDROID__
+		private protected override void UpdateSemanticInfoCore(SemanticInfoRequest request)
+		{
+			base.UpdateSemanticInfoCore(request);
+
+
+			foreach (var gesture in this.GestureRecognizers)
+			{
+				//Accessibility can't handle Tap Recognizers with > 1 tap
+				if (gesture is TapGestureRecognizer tgr && tgr.NumberOfTapsRequired == 1 && request.info != null)
+				{
+					request.info.AddAction(AndroidX.Core.View.Accessibility.AccessibilityNodeInfoCompat.AccessibilityActionCompat.ActionClick);
+				}
+			}
+
+		}
+#endif
+
 	}
 }
