+++ src/Core/src/Core/ISemantic.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Maui
+{
+	public interface ISemantic
+	{
+		/// <summary>
+		/// Adds semantics to every FrameworkElement for accessibility
+		/// </summary>
+		Semantics Semantics { get; }
+
+		void UpdateSemanticInfo(SemanticInfoRequest request);
+	}
+}
