+++ src/Core/src/Platform/Android/MauiAccessibilityDelegate.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using AndroidX.Core.View;
+using AndroidX.Core.View.Accessibility;
+using Microsoft.Maui.Handlers;
+using NativeView = Android.Views.View;
+
+namespace Microsoft.Maui
+{
+	class MauiAccessibilityDelegate : AccessibilityDelegateCompat
+	{
+		public ViewHandler? Handler { get; set; }
+
+		public MauiAccessibilityDelegate()
+		{
+		}
+
+		public override void OnInitializeAccessibilityNodeInfo(NativeView? host, AccessibilityNodeInfoCompat? info)
+		{
+			base.OnInitializeAccessibilityNodeInfo(host, info);
+
+			if (Handler?.NativeView is NativeView nativeView && Handler?.VirtualView != null)
+				nativeView.UpdateSemanticNodeInfo(Handler.VirtualView, info);
+		}
+	}
+}
