+++ src/Core/src/Platform/iOS/ViewExtensions.cs
@@ -100,26 +100,6 @@ namespace Microsoft.Maui
 				wrapper.Clip = view.Clip;
 		}
 
-		public static void UpdateSemantics(this UIView nativeView, IView view)
-		{
-			var semantics = view.Semantics;
-
-			if (semantics == null)
-				return;
-
-			nativeView.AccessibilityLabel = semantics.Description;
-			nativeView.AccessibilityHint = semantics.Hint;
-
-			// UIControl elements automatically have IsAccessibilityElement set to true
-			if (nativeView is not UIControl && (!string.IsNullOrWhiteSpace(semantics.Hint) || !string.IsNullOrWhiteSpace(semantics.Description)))
-				nativeView.IsAccessibilityElement = true;
-
-			if (semantics.IsHeading)
-				nativeView.AccessibilityTraits |= UIAccessibilityTrait.Header;
-			else
-				nativeView.AccessibilityTraits &= ~UIAccessibilityTrait.Header;
-		}
-
 		public static T? FindDescendantView<T>(this UIView view) where T : UIView
 		{
 			var queue = new Queue<UIView>();
